MusicCaps dataset
The dataset consisting of youtube ids of the sounds with time stamps and aspect_list. In this notebook, I added a column "midi" which initially consists of the youtube url derived from the youtube id. Subsequently, I tried to use the libraries - pytube and pydub to convert those urls to midi files as required by the training model.